Extract test utilities into separate crate
Pull fixscript into its own crate
so that it can be imported from integration tests.
