+++ testtools/src/fixscript.rs
@@ -1,0 +1,304 @@
+//! Builds directories from shell scripts.
+//!
+//! With fixscript, you can write a shell script that produces a directory,
+//! and have it recorded in the repository for later reuse.
+//!
+//! Each "fixture" has two pieces:
+//!
+//! - a shell script
+//! - an archive that holds the directory generated by the script
+//!
+//! The module ensures that these two are in-sync:
+//! if the archive doesn't exist or becomes out-of-date,
+//! fixscript replaces it with a new one that matches the current script's contents.
+//!
+//! fixscript specifically aims to address building test Git repositories,
+//! so it sets a number of Git-specific environment variables
+//! when running the shell scripts.
+//!
+//! This approach is heavily inspired by Gitoxide's integration testing system.
+
+use anyhow::{bail, Context, Result};
+use sha2::Digest;
+use std::fmt::Debug;
+use std::result::Result as StdResult;
+use std::{borrow::Cow, env, ffi, fs, io, path, process};
+
+/// Version of the format used by fixscript in generated archives.
+/// This helps ensure that we can change the format later.
+const VERSION: &str = "1";
+
+/// Handle to a directory generated by fixscript.
+///
+/// The directory will be deleted automatically when this is dropped.
+pub struct Fixture<'a> {
+    tempdir: tempfile::TempDir,
+    sha: Vec<u8>,
+    version: Cow<'a, str>,
+}
+
+impl<'a> Debug for Fixture<'a> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("Fixture")
+            .field("tempdir", &self.tempdir.path())
+            .field("sha", &format!("{:0x?}", &self.sha))
+            .field("version", &self.version)
+            .finish()
+    }
+}
+
+impl Fixture<'_> {
+    /// Reports the directory generated by fixscript.
+    pub fn dir(&self) -> &path::Path {
+        self.tempdir.path()
+    }
+
+    fn open_archive(path: &path::Path) -> Result<Self> {
+        let tempdir = tempfile::tempdir().context("cannot create temporary directory")?;
+
+        {
+            let f = fs::File::open(&path).context("cannot open archive")?;
+            let xz_dec = xz2::read::XzDecoder::new(f);
+            tar::Archive::new(xz_dec)
+                .unpack(tempdir.path())
+                .context("cannot extract archive")?;
+        }
+
+        let sha = fs::read(tempdir.path().join("SHA256")).context("failed to read fixture hash")?;
+        let version = fs::read_to_string(tempdir.path().join("VERSION"))
+            .context("failed to read fixture version")?;
+
+        // TODO: hash validation here and return error
+
+        Ok(Fixture {
+            tempdir,
+            sha,
+            version: Cow::Owned(version),
+        })
+    }
+}
+
+/// getenv is the default implementation of Group.getenv.
+fn getenv(s: &str) -> StdResult<String, env::VarError> {
+    env::var(s)
+}
+
+/// Group is a group of fixtures rooted in the same directory.
+/// This exists mostly for testing -- externally, the DEFAULT_GROUP is used.
+pub struct Group {
+    dir: path::PathBuf,
+    getenv: fn(&str) -> StdResult<String, env::VarError>,
+}
+
+impl Group {
+    pub fn new<P: AsRef<path::Path>>(p: P) -> Self {
+        Self {
+            dir: p.as_ref().to_path_buf(),
+            getenv,
+        }
+    }
+
+    /// Reports the absolute path to the given file/folder inside the fixtures/
+    /// directory.
+    fn fixture_path<P: AsRef<path::Path>>(&self, p: P) -> path::PathBuf {
+        self.dir.join(p)
+    }
+
+    /// Opens and returns a Fixture for the script at `script_path`,
+    /// ensuring that the archive for the script exists and is up to date.
+    ///
+    /// The path must be relative to the "fixtures" directory of the repository.
+    pub fn open<P: AsRef<path::Path>>(&self, script_name: P) -> Result<Fixture> {
+        let script_path = self.fixture_path(&script_name);
+        let archive_path = script_path.with_extension("tar.xz");
+
+        let script_sha = hash_file(&script_path).context("cannot hash fixture script")?;
+
+        if archive_path.exists() {
+            let fix = Fixture::open_archive(&archive_path).with_context(|| {
+                format!("unable to open fixture archive {}", archive_path.display())
+            })?;
+            if fix.sha == script_sha && fix.version == VERSION {
+                return Ok(fix);
+            }
+            eprintln!("archive {} is outdated", archive_path.display());
+        }
+
+        // Fail if the archive wasn't checked in and we're already in CI.
+        // GitHub workflows always sets "CI=true".
+        if let Ok(ci) = (self.getenv)("CI") {
+            if ci.as_str() == "true" {
+                eprintln!("cannot generate archive {} in CI", archive_path.display());
+                eprintln!("please run the test locally and check the archive in.");
+                bail!(
+                    "archive {} is outdated or does not exist",
+                    archive_path.display()
+                );
+            }
+        }
+
+        // TODO: Move new fixture creation into impl Fixture.
+
+        let tempdir = tempfile::tempdir().context("create temporary directory")?;
+        fs::write(tempdir.path().join("SHA256"), &script_sha).context("write SHA file")?;
+        fs::write(tempdir.path().join("VERSION"), VERSION.as_bytes())
+            .context("write VERSION file")?;
+
+        let script_path_abs = script_path.canonicalize()?;
+
+        let mut new_path = ffi::OsString::new();
+        if let Ok(bin) = self.fixture_path("bin").canonicalize() {
+            new_path.push(bin);
+        }
+        new_path.push(":");
+        new_path.push((self.getenv)("PATH")?);
+
+        let status = process::Command::new("bash")
+            .args(&["-euo", "pipefail"]) // disallow failures
+            .arg(&script_path_abs)
+            .current_dir(tempdir.path())
+            .env_remove("GIT_DIR")
+            .env("GIT_AUTHOR_DATE", "2000-01-01 00:00:00 +0000")
+            .env("GIT_AUTHOR_EMAIL", "author@example.com")
+            .env("GIT_AUTHOR_NAME", "author")
+            .env("GIT_COMMITTER_DATE", "2000-01-02 00:00:00 +0000")
+            .env("GIT_COMMITTER_EMAIL", "committer@example.com")
+            .env("GIT_COMMITTER_NAME", "committer")
+            .env("PATH", &new_path)
+            .status()
+            .with_context(|| format!("cannot run {}", script_path_abs.display()))?;
+        if !status.success() {
+            bail!("fixture {} failed", script_name.as_ref().display());
+        }
+
+        {
+            let f = fs::File::create(&archive_path)
+                .with_context(|| format!("create {}", archive_path.display()))?;
+
+            let out = xz2::write::XzEncoder::new(f, 3);
+            let mut ar = tar::Builder::new(out);
+            ar.append_dir_all(".", tempdir.path())?;
+        }
+
+        Ok(Fixture {
+            tempdir,
+            sha: script_sha,
+            version: Cow::Borrowed(VERSION),
+        })
+    }
+}
+
+/// Hashes the file at the given location.
+fn hash_file(p: &path::Path) -> Result<Vec<u8>> {
+    let mut f = fs::File::open(&p).context("cannot open file")?;
+    let mut hasher = sha2::Sha256::new();
+    io::copy(&mut f, &mut hasher).context("cannot hash file contents")?;
+    Ok(hasher.finalize().to_vec())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use anyhow::Result;
+    use tempfile;
+
+    fn getenv(s: &str) -> StdResult<String, env::VarError> {
+        match s {
+            "CI" => Err(env::VarError::NotPresent),
+            _ => env::var(s),
+        }
+    }
+
+    #[test]
+    fn open_end_to_end() -> Result<()> {
+        let tempdir = tempfile::tempdir()?;
+        let fixtures = tempdir.path();
+        let grp = Group {
+            dir: fixtures.to_path_buf(),
+            getenv,
+        };
+
+        let assert_fixture_contents = |want: &str| -> Result<()> {
+            let fix = grp.open("test.sh").context("cannot open fixture")?;
+            let got =
+                fs::read_to_string(fix.dir().join("bar")).context("cannot read generated file")?;
+            assert_eq!(want, got, "contents of generated file did not match");
+
+            Ok(())
+        };
+
+        // Create a fixture for the first time.
+        fs::write(fixtures.join("test.sh"), "echo foo > bar").context("cannot write test file")?;
+        assert_fixture_contents("foo\n").context("could not create initial archive from script")?;
+
+        // Verify archive exists and read from it.
+        let archive_hash =
+            hash_file(&fixtures.join("test.tar.xz")).context("cannot hash generated archive")?;
+        assert_fixture_contents("foo\n").context("could not open existing archive")?;
+
+        // Invalidate the archive and reopen.
+        fs::write(fixtures.join("test.sh"), "echo 'baz qux' > bar")
+            .context("cannot overwrite test file")?;
+        assert_fixture_contents("baz qux\n").context("could not recreate outdated archive")?;
+        let new_archive_hash =
+            hash_file(&fixtures.join("test.tar.xz")).context("cannot hash generated archive")?;
+
+        assert_ne!(
+            archive_hash, new_archive_hash,
+            "generated archive should be updated when fixture changes"
+        );
+
+        Ok(())
+    }
+
+    #[test]
+    fn bad_fixture_script() -> Result<()> {
+        let tempdir = tempfile::tempdir()?;
+        let fixtures = tempdir.path();
+        let grp = Group {
+            dir: fixtures.to_path_buf(),
+            getenv,
+        };
+
+        fs::write(fixtures.join("test.sh"), "false")?;
+        let err = grp
+            .open("test.sh")
+            .expect_err("fixture execution should fail");
+        assert!(
+            format!("{}", err).contains("fixture test.sh failed"),
+            "unexpected message: {}",
+            err
+        );
+
+        Ok(())
+    }
+
+    fn getenv_ci(s: &str) -> StdResult<String, env::VarError> {
+        match s {
+            "CI" => Ok("true".to_string()),
+            _ => env::var(s),
+        }
+    }
+
+    #[test]
+    fn new_archives_not_allowed_in_ci() -> Result<()> {
+        let tempdir = tempfile::tempdir()?;
+        let fixtures = tempdir.path();
+        let grp = Group {
+            dir: fixtures.to_path_buf(),
+            getenv: getenv_ci,
+        };
+
+        fs::write(fixtures.join("test.sh"), "echo foo > bar")?;
+        let err = grp
+            .open("test.sh")
+            .expect_err("fixture execution should fail");
+        assert!(
+            format!("{}", err).contains("outdated or does not exist"),
+            "unexpected message: {}",
+            err
+        );
+
+        Ok(())
+    }
+}
